requirements.txt
crewai
fastapi
uvicorn
pydantic
python-dotenv
